refactor: adaptor
